+++ js/main.js
@@ -1,106 +1,3 @@
-const PHOTO_DESCRIPTIONS = [
-  'поле',
-  'табличка',
-  'озеро',
-  'девушка',
-  'шоколад',
-  'черная машина',
-  'клубника',
-  'стакан',
-  'самолёт',
-  'ботинки',
-  'песок',
-  'белая машина',
-  'мясо',
-  'кот',
-  'валенки',
-  'небо',
-  'хор',
-  'кирпич',
-  'фонарик',
-  'пальма',
-  'вилка',
-  'море',
-  'краб',
-  'концерт',
-  'бегемот',
-];
+import {createDescriptionImages} from './data.js';
 
-const COMMENT_MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const USER_NAMES = [
-  'Иван',
-  'Андрей',
-  'Дима',
-  'Анна',
-  'Ольга',
-  'Мария',
-  'Анастасия',
-  'Евгений',
-  'Геннадий',
-  'Леопольд',
-];
-
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomElement = (elements) =>
-  elements[getRandomInteger(0, elements.length - 1)];
-
-const createIdGenerator = (min, max, uniqueness = false) => {
-  const storage = [];
-
-  return () => {
-    let randomValue = getRandomInteger(min, max);
-
-    if (uniqueness) {
-      while (storage.includes(randomValue)) {
-        randomValue = getRandomInteger(min, max);
-      }
-      storage.push(randomValue);
-    }
-    return randomValue;
-  };
-};
-
-const getPhotoId = createIdGenerator(1, 25, true);
-const getCommentId = createIdGenerator(1, 1_000, true);
-
-const generateDescription = () => getRandomElement(PHOTO_DESCRIPTIONS);
-
-const generateMessage = () =>
-  Array.from({ length: getRandomInteger(1, 2) }, () =>
-    getRandomElement(COMMENT_MESSAGES)
-  ).join(' ');
-
-const generateComments = () => ({
-  id: getCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: generateMessage(),
-  name: getRandomElement(USER_NAMES),
-});
-
-const createDescriptionImage = () => {
-  const id = getPhotoId();
-
-  return {
-    id,
-    url: `photos/${id}.jpg`,
-    description: generateDescription(),
-    likes: getRandomInteger(15, 200),
-    comments: Array.from({ length: getRandomInteger(0, 30) }, generateComments),
-  };
-};
-
-export { createDescriptionImage };
+createDescriptionImages();
